cleaned up the carbon_update branch

This ensures that all components are working properly and that the documentation is complete with respect to the protected area and carbon updates.

+++ src/read_country_fao.c
@@ -88,7 +88,7 @@ int read_country_fao(args_struct in_args, rinfo_struct *raster_info) {
 	}
 	
 	// read the data
-	num_read = fread(country_fao, insize, ncells, fpin);
+	num_read = (int) fread(country_fao, insize, ncells, fpin);
 	fclose(fpin);
 	if(num_read != ncells)
 	{
